+++ quad_mesh_simplify/c/valid_pairs.h
@@ -2,4 +2,4 @@
 #include "mesh.h"
 #include "sparse_mat.h"
 
-Array2D_uint* compute_valid_pairs(Mesh* mesh, SparseMat* edges, unsigned int threshold);
+Array2D_uint* compute_valid_pairs(Mesh* mesh, SparseMat* edges, double threshold);
